fixed pass bugs
