+++ components/Login.js
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { useState} from 'react';
 import {StyleSheet, Button, TextInput, Text, View, Alert} from 'react-native';
 
 
@@ -6,56 +6,9 @@ import {StyleSheet, Button, TextInput, Text, View, Alert} from 'react-native';
 export default function Login(props) {
 
   
-  const [errMsg, setErrMsg] = useState('');
+ 
   const [userText, setUser] = useState('');
   const [passText, setPass] = useState('');
-
-  function checkPassword(pass) {
-    setErrMsg('');
-    minRegex = new RegExp("(?=.{6,})");
-    lalpha = new RegExp("(?=.*[a-z])");
-    ualpha = new RegExp("(?=.*[A-Z])");
-    num = new RegExp("(?=.*[0-9])");
-    special = new RegExp("(?=.*[!@#$%^&*])");
-    if (pass ==null || pass === '') {
-        setErrMsg('Password is required.');
-    }
-    else if (!minRegex.test(pass)) {
-        setErrMsg('Password must be longer than 6 characaters.');
-    }
-    else if (!lalpha.test(pass)) {
-        setErrMsg('Password must contain at least one lowercase alphabetical character.');
-    }
-    else if (!ualpha.test(pass)) {
-        setErrMsg('Password must contain at least one uppercase alphabetical character.');
-    }
-    else if (!num.test(pass)) {
-        setErrMsg('Password must contain at least one numeric character.');
-    }
-    else if (!special.test(pass)) {
-        setErrMsg('Password must contain at least one special character.');
-    }
-    if (errMsg.length > 0) {
-        Alert.alert(
-            "Error",
-            errMsg,
-            [
-              {
-                text: "Cancel",
-                onPress: () => console.log("Cancel Pressed"),
-                style: "cancel"
-              },
-              { text: "OK", onPress: () => console.log("OK Pressed") }
-            ],
-            { cancelable: false }
-          );
-    }
-    
-  }
-
-  
-  
-
 
   return (
     
@@ -81,7 +34,6 @@ export default function Login(props) {
       <View style={styles.buttonContainer}>
         <Button
             title = "SIGN IN"
-            onPress={() => checkPassword(passText)}
         />
         <Button
             title = "Login with Facebook"
